Move functions to own package
