+++ h2/src/main/org/h2/expression/IntervalOperation.java
@@ -20,10 +20,10 @@ import java.math.BigInteger;
 import org.h2.api.ErrorCode;
 import org.h2.api.IntervalQualifier;
 import org.h2.engine.Session;
+import org.h2.expression.function.DateTimeFunctions;
 import org.h2.message.DbException;
 import org.h2.table.ColumnResolver;
 import org.h2.table.TableFilter;
-import org.h2.util.DateTimeFunctions;
 import org.h2.util.IntervalUtils;
 import org.h2.value.DataType;
 import org.h2.value.Value;
