+++ h2/src/main/org/h2/expression/function/FunctionCall.java
@@ -3,9 +3,10 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
-package org.h2.expression;
+package org.h2.expression.function;
 
 import org.h2.engine.Session;
+import org.h2.expression.Expression;
 import org.h2.value.ValueResultSet;
 
 /**
