+++ h2/src/main/org/h2/expression/function/FunctionInfo.java
@@ -3,7 +3,7 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
-package org.h2.expression;
+package org.h2.expression.function;
 
 /**
  * This class contains information about a built-in function.
