+++ h2/src/main/org/h2/expression/function/ToDateParser.java
@@ -3,7 +3,7 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: Daniel Gredler
  */
-package org.h2.util;
+package org.h2.expression.function;
 
 import static java.lang.String.format;
 
@@ -12,6 +12,7 @@ import java.util.GregorianCalendar;
 import java.util.List;
 import java.util.TimeZone;
 
+import org.h2.util.DateTimeUtils;
 import org.h2.value.ValueTimestamp;
 import org.h2.value.ValueTimestampTimeZone;
