+++ h2/src/main/org/h2/mode/FunctionsBase.java
@@ -8,8 +8,8 @@ package org.h2.mode;
 import java.util.HashMap;
 
 import org.h2.engine.Database;
-import org.h2.expression.Function;
-import org.h2.expression.FunctionInfo;
+import org.h2.expression.function.Function;
+import org.h2.expression.function.FunctionInfo;
 
 /**
  * Base class for mode-specific functions.
