+++ h2/src/main/org/h2/mode/FunctionsMySQL.java
@@ -14,9 +14,9 @@ import org.h2.api.ErrorCode;
 import org.h2.engine.Database;
 import org.h2.engine.Session;
 import org.h2.expression.Expression;
-import org.h2.expression.Function;
-import org.h2.expression.FunctionInfo;
 import org.h2.expression.ValueExpression;
+import org.h2.expression.function.Function;
+import org.h2.expression.function.FunctionInfo;
 import org.h2.message.DbException;
 import org.h2.util.StringUtils;
 import org.h2.value.DataType;
