+++ h2/src/main/org/h2/table/FunctionTable.java
@@ -9,8 +9,8 @@ import java.util.ArrayList;
 import org.h2.api.ErrorCode;
 import org.h2.engine.Session;
 import org.h2.expression.Expression;
-import org.h2.expression.FunctionCall;
-import org.h2.expression.TableFunction;
+import org.h2.expression.function.FunctionCall;
+import org.h2.expression.function.TableFunction;
 import org.h2.index.FunctionIndex;
 import org.h2.index.Index;
 import org.h2.index.IndexType;
