+++ h2/src/test/org/h2/test/db/TestFunctions.java
@@ -46,6 +46,8 @@ import org.h2.api.Aggregate;
 import org.h2.api.AggregateFunction;
 import org.h2.api.ErrorCode;
 import org.h2.engine.Constants;
+import org.h2.expression.function.ToDateParser;
+import org.h2.expression.function.ToChar.Capitalization;
 import org.h2.message.DbException;
 import org.h2.store.fs.FileUtils;
 import org.h2.test.TestBase;
@@ -55,8 +57,6 @@ import org.h2.tools.SimpleResultSet;
 import org.h2.util.DateTimeUtils;
 import org.h2.util.IOUtils;
 import org.h2.util.StringUtils;
-import org.h2.util.ToChar.Capitalization;
-import org.h2.util.ToDateParser;
 import org.h2.value.Value;
 import org.h2.value.ValueTimestamp;
 import org.h2.value.ValueTimestampTimeZone;
